+++ src/Api.php
@@ -4,11 +4,17 @@ namespace Razorpay\Api;
 
 class Api
 {
-    public static $baseUrl = "https://api.razorpay.com/v1/";
+    protected static $baseUrl = 'https://api.razorpay.com/v1/';
 
-    public static $key = null;
+    protected static $key = null;
 
-    public static $secret = null;
+    protected static $secret = null;
+
+    /*
+     * App info is to store the Plugin/integration
+     * information
+     */
+    public static $appsDetails = array();
 
     const VERSION = '1.2.9';
 
@@ -16,34 +22,71 @@ class Api
      * @param string $key
      * @param string $secret
      */
-    function __construct($key, $secret)
+    public function __construct($key, $secret)
     {
         self::$key = $key;
         self::$secret = $secret;
-
-        // Add the version to all HTTP Requests
-        Request::addHeader('User-Agent', "Razorpay-PHP/".self::VERSION);
     }
 
     /*
      *  Set Headers
      *
      */
-    function setHeader($header, $value)
+    public function setHeader($header, $value)
     {
         Request::addHeader($header, $value);
+    }
+
+    public function setAppDetails($title, $version = null)
+    {
+        $app = array(
+            'title' => $title,
+            'version' => $version
+        );
+
+        array_push(self::$appsDetails, $app);
+    }
+
+    public function getAppsDetails()
+    {
+        return self::$appsDetails;
+    }
+
+    public function setBaseUrl($baseUrl)
+    {
+        self::$baseUrl = $baseUrl;
     }
 
     /**
      * @param string $name
      * @return mixed
      */
-    function __get($name)
+    public function __get($name)
     {
         $className = __NAMESPACE__.'\\'.ucwords($name);
 
         $entity = new $className();
 
         return $entity;
+    }
+
+    public static function getBaseUrl()
+    {
+        return self::$baseUrl;
+    }
+
+    public static function getKey()
+    {
+        return self::$key;
+    }
+
+    public static function getSecret()
+    {
+        return self::$secret;
+    }
+
+    public static function getFullUrl($relativeUrl)
+    {
+        return self::getBaseUrl() . $relativeUrl;
     }
 }