+++ src/Payment.php
@@ -22,9 +22,11 @@ class Payment extends Entity
      */
     public function refund($attributes = array())
     {
-        $relativeUrl = $this->getEntityUrl() . $this->id . '/refund';
+        $refund = new Refund;
 
-        return $this->request('POST', $relativeUrl, $attributes);
+        $attributes = array_merge($attributes, array('payment_id' => $this->id));
+
+        return $refund->create($attributes);
     }
 
     /**
@@ -39,10 +41,10 @@ class Payment extends Entity
 
     public function refunds()
     {
-        $refund = new Refund();
+        $refund = new Refund;
 
-        $refund['payment_id'] = $this->id;
+        $options = array('payment_id' => $this->id);
 
-        return $refund;
+        return $refund->all($options);
     }
 }